Fix the wrong container id in SBE numeric effecter
This commit solves the issue caused by #319 where we
persisted the HB entity association PDRs before we normalise
them. This lead to assigning the wrong container ids for
the SBE numeric effecter PDRs created.
With this fix, we assign the normalised container ids to the
numeric effecter PDRs.

This commit also removes the stale code to update the container
ids of proc LEDs. The JSON file is already updated with the
correct conatiner id and this piece of code to update it is
not needed.

Fixes: PE00DXJZ
