+++ oem/ibm/libpldm/pdr_oem_ibm.h
@@ -6,6 +6,7 @@ extern "C" {
 #endif
 
 #include "pdr_data.h"
+#include "platform.h"
 #include <assert.h>
 #include <stddef.h>
 #include <stdint.h>
@@ -19,6 +20,21 @@ extern "C" {
  */
 pldm_pdr_record *pldm_pdr_find_last_local_record(const pldm_pdr *repo);
 
+/** @brief method to check if the record handle is within the HostBoot range
+ *  or not
+ *
+ *  @param[in] record_handle - record handle of the pdr
+ */
+bool isHBRange(const uint32_t record_handle);
+
+/** @brief find the container ID of the contained entity
+ *
+ *  @param[in] repo - opaque pointer acting as a PDR repo handle
+ *  @param[in] entityType - entity type
+ *  @param[in] entityInstance - instance of the entity
+ */
+uint16_t pldm_find_container_id(const pldm_pdr *repo, uint16_t entityType,
+				uint16_t entityInstance);
 #ifdef __cplusplus
 }
 #endif
