+++ oem/ibm/libpldmresponder/oem_ibm_handler.cpp
@@ -1576,129 +1576,6 @@ uint8_t pldm::responder::oem_ibm_platform::Handler::fetchRealSAIStatus()
 }
 
 void pldm::responder::oem_ibm_platform::Handler::
-    updateIdentifyEffecterAndSensorPDRs()
-{
-    pldm::pdr::EntityType entityType = PLDM_ENTITY_PROC;
-
-    auto stateEffecterPDRs = pldm::utils::findStateEffecterPDR(
-        mctp_eid, entityType,
-        static_cast<uint16_t>(PLDM_STATE_SET_IDENTIFY_STATE), pdrRepo);
-
-    if (stateEffecterPDRs.empty())
-    {
-        std::cerr << "The State effecter PDR cannot be found, entityType = "
-                  << entityType << std::endl;
-        return;
-    }
-
-    for (auto& effecterPDR : stateEffecterPDRs)
-    {
-        for (auto& [key, value] : instanceMap)
-        {
-            auto instNumber = (value.dcmId * 2) + value.procId;
-            auto pdr =
-                reinterpret_cast<pldm_state_effecter_pdr*>(effecterPDR.data());
-            if (instNumber == ((pdr->entity_instance) - 1))
-            {
-                uint16_t newContainerID = pldm_find_container_id(
-                    pdrRepo, PLDM_ENTITY_PROC_MODULE, value.dcmId);
-                pldm_change_container_id_of_effecter(pdrRepo, pdr->effecter_id,
-                                                     newContainerID);
-                pldm_change_instance_number_of_effecter(
-                    pdrRepo, pdr->effecter_id, (instNumber % 2));
-                break;
-            }
-        }
-    }
-
-    auto stateSensorIdentifyPDRs = findStateSensorPDR(
-        mctp_eid, entityType,
-        static_cast<uint16_t>(PLDM_STATE_SET_IDENTIFY_STATE), pdrRepo);
-    for (auto& identifySensorPDR : stateSensorIdentifyPDRs)
-    {
-        for (auto& [key, value] : instanceMap)
-        {
-            auto instanceNumber = (value.dcmId * 2) + value.procId;
-            auto pdr = reinterpret_cast<pldm_state_sensor_pdr*>(
-                identifySensorPDR.data());
-            if (instanceNumber == ((pdr->entity_instance) - 1))
-            {
-                uint16_t newContainerID = pldm_find_container_id(
-                    pdrRepo, PLDM_ENTITY_PROC_MODULE, value.dcmId);
-                pldm_change_container_id_of_sensor(pdrRepo, pdr->sensor_id,
-                                                   newContainerID);
-                pldm_change_instance_number_of_sensor(pdrRepo, pdr->sensor_id,
-                                                      (instanceNumber % 2));
-                break;
-            }
-        }
-    }
-}
-
-void pldm::responder::oem_ibm_platform::Handler::
-    updateFaultEffecterAndSensorPDRs()
-{
-    pldm::pdr::EntityType entityType = PLDM_ENTITY_PROC;
-    auto stateEffecterFaultPDRs = pldm::utils::findStateEffecterPDR(
-        mctp_eid, entityType,
-        static_cast<uint16_t>(PLDM_STATE_SET_OPERATIONAL_FAULT_STATUS),
-        pdrRepo);
-    for (auto& faultEffecterPDR : stateEffecterFaultPDRs)
-    {
-        for (auto& [key, value] : instanceMap)
-        {
-            auto instanceNumber = (value.dcmId * 2) + value.procId;
-            auto pdr = reinterpret_cast<pldm_state_effecter_pdr*>(
-                faultEffecterPDR.data());
-            if (instanceNumber == (pdr->entity_instance) - 1)
-            {
-                uint16_t newContainerID = pldm_find_container_id(
-                    pdrRepo, PLDM_ENTITY_PROC_MODULE, value.dcmId);
-                pldm_change_container_id_of_effecter(pdrRepo, pdr->effecter_id,
-                                                     newContainerID);
-                pldm_change_instance_number_of_effecter(
-                    pdrRepo, pdr->effecter_id, (instanceNumber % 2));
-                break;
-            }
-        }
-    }
-
-    auto stateSensorFaultPDRs = findStateSensorPDR(
-        mctp_eid, entityType,
-        static_cast<uint16_t>(PLDM_STATE_SET_OPERATIONAL_FAULT_STATUS),
-        pdrRepo);
-    for (auto& faultSensorPDR : stateSensorFaultPDRs)
-    {
-        for (auto& [key, value] : instanceMap)
-        {
-            auto instanceNumber = (value.dcmId * 2) + value.procId;
-            auto pdr =
-                reinterpret_cast<pldm_state_sensor_pdr*>(faultSensorPDR.data());
-            if (instanceNumber == (pdr->entity_instance) - 1)
-            {
-                uint16_t newContainerID = pldm_find_container_id(
-                    pdrRepo, PLDM_ENTITY_PROC_MODULE, value.dcmId);
-                pldm_change_container_id_of_sensor(pdrRepo, pdr->sensor_id,
-                                                   newContainerID);
-                pldm_change_instance_number_of_sensor(pdrRepo, pdr->sensor_id,
-                                                      (instanceNumber % 2));
-                break;
-            }
-        }
-    }
-}
-
-void pldm::responder::oem_ibm_platform::Handler::updateContainerIDofProcLed()
-{
-    if (update)
-    {
-        updateIdentifyEffecterAndSensorPDRs();
-        updateFaultEffecterAndSensorPDRs();
-        update = false;
-    }
-}
-
-void pldm::responder::oem_ibm_platform::Handler::
     processPowerCycleOffSoftGraceful()
 {
     std::cerr << "Received soft graceful power cycle request" << std::endl;
@@ -1922,17 +1799,6 @@ void pldm::responder::oem_ibm_platform::Handler::triggerHostEffecter(
         }
     }
 }
-
-bool pldm::responder::oem_ibm_platform::Handler::isHBRange(
-    const uint32_t& record_handle)
-{
-    if (record_handle >= 0x01000000 && record_handle < 0x01FFFFFF)
-    {
-        return true;
-    }
-    return false;
-}
-
 } // namespace oem_ibm_platform
 } // namespace responder
 } // namespace pldm
